refactor(matrices): move add, mul, extract to RepMatrix
